+++ crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh/src/witness/proofs.rs
@@ -1,21 +1,28 @@
 /*!
 *   A webvh that has witnessing enabled requires a proof file containing each witness proof
+*
+*   When saving or serializing the Witness Proofs, you should run `optimise_records` first
+*   THis will ensure that previous witness proof records have been removed
 */
-use affinidi_data_integrity::DataIntegrityProof;
-use serde::{Deserialize, Serialize};
+use std::fs::File;
 
 use crate::DIDWebVHError;
+use affinidi_data_integrity::DataIntegrityProof;
+use ahash::{HashMap, HashMapExt};
+use serde::{Deserialize, Serialize};
+use tracing::warn;
 
 // *********************************************************
 // Witness Proof File
 // *********************************************************
 
 /// Array of WitnessProofs for each Witness Proof
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct WitnessProofCollection(pub Vec<WitnessProof>);
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+pub struct WitnessProofCollection(Vec<WitnessProof>);
 
 /// Record of each LogEntry that requires witnessing
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
 pub struct WitnessProof {
     /// versionId of the DID Log Entry to which witness proofs apply.
     pub version_id: String,
@@ -59,6 +66,24 @@ impl WitnessProofCollection {
             .map_or(0, |p| p.proof.len())
     }
 
+    /// Load existing proofs from a file
+    pub fn read_from_file(file_path: &str) -> Result<Self, DIDWebVHError> {
+        let file = File::open(file_path).map_err(|e| {
+            DIDWebVHError::WitnessProofError(format!(
+                "Couldn't open Witness Proofs file ({}): {}",
+                file_path, e
+            ))
+        })?;
+        let proofs: WitnessProofCollection = serde_json::from_reader(file).map_err(|e| {
+            DIDWebVHError::WitnessProofError(format!(
+                "Couldn't deserialize Witness Proofs Data from file ({}): {}",
+                file_path, e
+            ))
+        })?;
+        Ok(proofs)
+    }
+
+    /// Save proofs to a file
     pub fn save_to_file(&self, file_path: &str) -> Result<(), DIDWebVHError> {
         let json_data = serde_json::to_string(self).map_err(|e| {
             DIDWebVHError::WitnessProofError(format!(
@@ -72,6 +97,67 @@ impl WitnessProofCollection {
                 file_path, e
             ))
         })?;
+        Ok(())
+    }
+
+    /// Get WitnessProof record for a given version_id
+    pub fn get_proofs(&self, version_id: &str) -> Option<&WitnessProof> {
+        self.0.iter().find(|p| p.version_id == version_id)
+    }
+
+    /// Runs through and removes wtiness proofs from earlier LogEntries that are not required
+    pub fn optimise_records(&mut self) -> Result<(), DIDWebVHError> {
+        let mut witness_versions: HashMap<String, usize> = HashMap::new();
+
+        // Map out which versions each witness is visible in
+        for v in &self.0 {
+            let Some((id, _)) = v.version_id.split_once('-') else {
+                return Err(DIDWebVHError::WitnessProofError(format!(
+                    "Invalid versionID ({}) in witness proofs! Expected n-hash, but missing n",
+                    v.version_id
+                )));
+            };
+            let Ok(id): Result<usize, _> = str::parse(id) else {
+                return Err(DIDWebVHError::WitnessProofError(format!(
+                    "Invalid versionID ({}) in witness proofs! expected n-hash, where n is a number!",
+                    v.version_id
+                )));
+            };
+
+            // Walk through each proof for this versionID
+            for p in &v.proof {
+                if let Some(record) = witness_versions.get_mut(&p.verification_method) {
+                    if &id > record {
+                        *record = id;
+                    }
+                } else {
+                    // Create new witness record
+                    witness_versions.insert(p.verification_method.clone(), id);
+                }
+            }
+        }
+
+        // Strip out older proofs as needed
+        self.0.retain_mut(|v| {
+            let Some((id, _)) = v.version_id.split_once('-') else {
+                warn!(
+                    "Invalid versionID ({}) in witness proofs! Expected n-hash, but missing n", v.version_id);
+                return false;
+            };
+            let Ok(id): Result<usize, _> = str::parse(id) else {
+                warn!(
+                    "Invalid versionID ({}) in witness proofs! expected n-hash, where n is a number!", v.version_id);
+            return false;
+            };
+
+            // Remove older proofs
+            v.proof
+                .retain(|p| &id >= witness_versions.get(&p.verification_method).unwrap_or(&0));
+            
+            // If version has no proofs, then remove it
+             !v.proof.is_empty()
+        });
+
         Ok(())
     }
 }